agent_server/requirements.txt
openai==2.32.0
fastapi[standard]==0.136.1
redis==7.0.0
pydantic-settings==2.14.0
PyJWT[crypto]==2.12.1
langchain==1.2.16
langchain-openai==1.2.1
langchain-mcp-adapters==0.2.2
langgraph==1.1.10
langgraph-checkpoint-redis==0.4.1
pydantic==2.13.3
openstacksdk==4.12.0